requirements.txt
djangorestframework==3.15.2
six==1.16.0
